+++ src/main/resources/templates/portion/portioningMain.html
@@ -137,7 +137,7 @@
     /*]]>*/
 </script>
 
-<script th:src="@{/js/portion/portioningMain.js}" de  fer></script>
+<script th:src="@{/js/portion/portioningMain.js}" defer></script>
 <script th:src="@{/js/portion/mainsort.js}" defer></script>
 <script th:src="@{/js/modalnoti.js}"></script>
 <script>
